jblog #finish

+++ mysite/mysite3/src/main/java/com/poscodx/mysite/service/GuestbookService.java
@@ -4,7 +4,9 @@ import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
+import com.poscodx.mysite.repository.GuestbookLogRepository;
 import com.poscodx.mysite.repository.GuestbookRepository;
 import com.poscodx.mysite.vo.GuestbookVo;
 
@@ -14,15 +16,32 @@ public class GuestbookService {
 	@Autowired
 	private GuestbookRepository guestbookRepository;
 	
+	@Autowired
+	private GuestbookLogRepository guestbookLogRepository;
+	
 	public List<GuestbookVo> getContentsList() {
 		return guestbookRepository.findAll();
 	}
 	
+	@Transactional
 	public void deleteContents(Long no, String password) {
-		guestbookRepository.deleteByNoAndPassword(no, password);
+		GuestbookVo vo = guestbookRepository.findByNo(no);
+		if(vo == null) {
+			return;
+		}
+		
+		int count = guestbookRepository.deleteByNoAndPassword(no, password);
+		if(count == 1) {
+			guestbookLogRepository.update(vo.getRegDate());
+		}
 	}
 	
+	@Transactional
 	public void addContents(GuestbookVo vo) {
+		int count = guestbookLogRepository.update();		
+		if(count == 0) {
+			guestbookLogRepository.insert();
+		}
 		guestbookRepository.insert(vo);
 	}
 }
